requirements.txt
#Requirements for libraraies used:
PyPDF2==1.26.0
langchain==0.1.0
python-docx==0.8.11
openpyxl==3.0.7
streamlit==0.89.0
python-dotenv==0.19.0
streamlit-chat==0.1.2
Pillow==8.3.1
asyncio==3.4.3
aioredis==1.3.1
fastapi==0.70.0
httpx==0.19.0
